Moved last methog out of Menu
	Also did a bit of Header cleanup
