+++ interface/src/devices/RealSense.cpp
@@ -8,6 +8,8 @@
 //  Distributed under the Apache License, Version 2.0.
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
 //
+
+#include "Application.h"
 #include "RealSense.h"
 #include "MainWindow.h"
 #include "Menu.h"
