+++ interface/src/ui/DialogsManager.cpp
@@ -10,7 +10,6 @@
 //
 
 #include <AccountManager.h>
-#include <Application.h>
 #include <MainWindow.h>
 #include <PathUtils.h>
 #include <XmppClient.h>
@@ -20,6 +19,7 @@
 #include "AttachmentsDialog.h"
 #include "BandwidthDialog.h"
 #include "CachesSizeDialog.h"
+#include "ChatWindow.h"
 #include "HMDToolsDialog.h"
 #include "LodToolsDialog.h"
 #include "LoginDialog.h"
