+++ interface/src/ui/LodToolsDialog.cpp
@@ -14,6 +14,7 @@
 #include <QDialogButtonBox>
 #include <QDoubleSpinBox>
 #include <QFormLayout>
+#include <QLabel>
 #include <QPalette>
 #include <QPushButton>
 #include <QSlider>
