+++ interface/src/ui/MetavoxelEditor.h
@@ -12,6 +12,7 @@
 #ifndef hifi_MetavoxelEditor_h
 #define hifi_MetavoxelEditor_h
 
+#include <QFormLayout>
 #include <QList>
 #include <QWidget>
